requirements.txt
flask==2.0.1
flask-login==0.5.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
boto3==1.26.137
python-dotenv==0.19.0
werkzeug==2.0.1
waitress==2.1.2
gunicorn==20.1.0
